@@ -2,12 +2,12 @@ package com.thing.quoter.fragment
 
 import android.content.Context
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v7.widget.LinearLayoutManager
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.thing.quoter.QuoterHelper
+import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.LinearLayoutManager
+import com.thing.quoter.helper.QuoterHelper
 
 import com.thing.quoter.R
 import com.thing.quoter.adapter.ProviderViewAdapter
