@@ -1,7 +1,6 @@
 package com.thing.quoter
 
 import android.Manifest
-import android.annotation.SuppressLint
 import android.content.Context
 import android.content.pm.PackageManager
 import android.graphics.Bitmap
@@ -9,15 +8,13 @@ import android.graphics.Canvas
 import android.net.Uri
 import android.os.Bundle
 import android.provider.MediaStore
-import android.support.v4.app.ActivityCompat
-import android.support.v4.content.ContextCompat
-import android.support.v7.app.AppCompatActivity
 import android.view.View
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
 import com.gunhansancar.changelanguageexample.helper.LocaleHelper
-import com.thing.quoter.model.Quote
 import java.io.ByteArrayOutputStream
 
-@SuppressLint("Registered")
 abstract class AppActivity : AppCompatActivity() {
 
     var isFirstTime = true
