@@ -1,4 +1,4 @@
-package com.thing.quoter;
+package com.thing.quoter.view;
 
 import android.content.Context;
 import android.view.GestureDetector;
